% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIGenrichr.R
\name{BIGenrichr}
\alias{BIGenrichr}
\title{Run EnrichR function on multiple Geneset databases on either gene list or modules of intertest}
\usage{
BIGenrichr(gene_list = NULL, ID = "SYMBOL", dbs = c("MSigDB_Hallmark_2020"))
}
\arguments{
\item{gene_list}{Here input either Genelist in SYMBOL or Ensembl ID format or input module object output from module function}

\item{ID}{gene id is either SYMBOL, ENSEMBL, ENTREZ}

\item{dbs}{Enter the geneset database you want to run Enrichr on the default is MSigDB Hallmark 2020 to check available database run enrichR::listEnrichrDb()}
}
\value{
list with enrichr result for each geneset database
}
\description{
Run EnrichR function on multiple Geneset databases on either gene list or modules of intertest
}
\examples{
#Get gene names for enrichment
gene_list <- list(HRV1 = names(example_gene_list[[1]]),
                  HRV2 = names(example_gene_list[[2]]))
BIGenrichr(gene_list, ID = "ENSEMBL")
}
\author{
Basilin Benson
}
