% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIGsea.R
\name{BIGsea}
\alias{BIGsea}
\title{BIGsea}
\usage{
BIGsea(
  gene_list,
  nperm = 1000,
  species = "human",
  ID = "SYMBOL",
  category = NULL,
  subcategory = NULL,
  db = NULL
)
}
\arguments{
\item{gene_list}{Named list object with named numeric vectors of gene symbols and logFC}

\item{nperm}{Numeric permutations for P-value calculations. Default is 1000}

\item{species}{Character string denoting species of interest. Default is human}

\item{ID}{Character string for type of ID used in gene_list. One of SYMBOL, ENTREZ, ENSEMBL. Default is SYMBOL}

\item{category}{Character string denoting Broad gene set database}

\item{subcategory}{Character string denoting Broad gene set sub-database \cr
\tabular{rrrrr}{
\strong{category} \tab    \strong{subcategory}\cr
C1  \tab      \cr
C2  \tab      CGP\cr
C2  \tab      CP\cr
C2  \tab      CP:BIOCARTA\cr
C2  \tab      CP:KEGG\cr
C2  \tab      CP:PID\cr
C2  \tab      CP:REACTOME\cr
C2  \tab      CP:WIKIPATHWAYS\cr
C3  \tab      \cr
C5  \tab      GO:BP\cr
C5  \tab      GO:CC\cr
C5  \tab      GO:MF\cr
C5  \tab      HPO\cr
C6  \tab      \cr
H   \tab      \cr
}}

\item{db}{If not using Broad databases, a data frame with gene ontology including category (gs_cat), subcategory (gs_subcat), gene set name (gs_name), and gene ID (gene_symbol, entrez_gene, or ensembl_gene as matches your gene_list names)}
}
\value{
Data frame of enrichments including pathway, significance, and leading edge genes
}
\description{
Run gene set enrichment analysis (GSEA) on multiple list using fgsea. More information about GSEA and geneset databases check https://www.gsea-msigdb.org/gsea/msigdb
}
\examples{
BIGsea(example_gene_list, category="H", ID="ENSEMBL")
BIGsea(example_gene_list, category="C2", subcategory="CP", ID="ENSEMBL")

#Use custom data base
db <- data.frame(module = c(rep("module1",10), rep("module2",10)),
                 symbol = sample(names(example_gene_list[[1]]), 20))
BIGsea(example_gene_list, ID="ENSEMBL", db=db)

}
