% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.human.db.full.R
\docType{data}
\name{ensembl.human.db.full}
\alias{ensembl.human.db.full}
\title{Background genes for flexEnrich, Ensembl IDs of all genes in the human genome}
\format{
A single-column matrix with Ensembl IDs
\describe{
\item{x}{character. 44454 gene IDs}
}
}
\source{
\url{https://useast.ensembl.org/info/data/biomart/index.html}
}
\usage{
ensembl.human.db.full
}
\description{
A single-column matrix of gene names
}
\keyword{datasets}
