% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez.human.db.full.R
\docType{data}
\name{entrez.human.db.full}
\alias{entrez.human.db.full}
\title{Background genes for flexEnrich, Entrez IDs of all genes in the human genome}
\format{
A single-column matrix with Entrez IDs
\describe{
\item{x}{character. 40755 gene IDs}
}
}
\source{
\url{https://useast.ensembl.org/info/data/biomart/index.html}
}
\usage{
entrez.human.db.full
}
\description{
A single-column matrix of gene names
}
\keyword{datasets}
