% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez.human.db.pc.R
\docType{data}
\name{entrez.human.db.pc}
\alias{entrez.human.db.pc}
\title{Background genes for flexEnrich, Entrez IDs of protein-coding genes in the human genome}
\format{
A single-column matrix with Entrez IDs
\describe{
\item{x}{character. 19383 gene IDs}
}
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/human/collections.jsp}
}
\usage{
entrez.human.db.pc
}
\description{
A single-column matrix of gene names
}
\keyword{datasets}
