% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgseaLabel2.R
\name{fgseaLabel2}
\alias{fgseaLabel2}
\title{Modify fgsea::fgseaLabel to use user input fold change estimates
Original authors Gennady Korotkevich, Vladimir Sukhov, Nikolay Budin, Alexey Sergushichev
See https://bioconductor.org/packages/release/bioc/html/fgsea.html}
\usage{
fgseaLabel2(
  pathways,
  dat,
  label,
  nperm,
  rand_var,
  rand_est,
  minSize = 1,
  maxSize = nrow(dat$E) - 1,
  nproc = 0,
  gseaParam = 1,
  BPPARAM = NULL,
  estimates = NULL,
  ...
)
}
\arguments{
\item{pathways}{List of gene sets to check.}

\item{dat}{Voom object with gene expression and sample metadata. Used in label randomization only'}

\item{label}{Numeric vector of labels for the correlation score of the same length as the number of columns in \code{mat}}

\item{nperm}{Number of permutations to do. Minimal possible nominal p-value is about 1/nperm}

\item{rand_var}{Character string specifying the variable to randomize in the label method}

\item{rand_est}{Data frame with random variable estimates. From prior run of BIGsea in the estimates slot}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{nproc}{If not equal to zero sets BPPARAM to use nproc workers (default = 0).}

\item{gseaParam}{GSEA parameter value, all gene-level statis are raised to the power of \code{gseaParam} before calculation of GSEA enrichment scores.}

\item{BPPARAM}{Parallelization parameter used in bplapply.
Can be used to specify cluster to run. If not initialized explicitly or
by setting \code{nproc} default value \code{bpparam()} is used.}

\item{estimates}{Named vector of numeric fold change estimates}

\item{...}{Additional parameters for kimma::kmFit needed if rand = "label"}
}
\value{
A table with GSEA results. Each row corresponds to a tested pathway.
}
\description{
Modify fgsea::fgseaLabel to use user input fold change estimates
Original authors Gennady Korotkevich, Vladimir Sukhov, Nikolay Budin, Alexey Sergushichev
See https://bioconductor.org/packages/release/bioc/html/fgsea.html
}
