% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexEnrich.R
\name{flexEnrich}
\alias{flexEnrich}
\title{Flexible hypergeometric enrichment}
\usage{
flexEnrich(
  gene_list = NULL,
  gene_df = NULL,
  ID = "SYMBOL",
  species = "human",
  collection = NULL,
  subcollection = NULL,
  db = NULL,
  custom_bg = NULL,
  protein_coding = TRUE,
  minOverlap = 1,
  minGeneSetSize = 10,
  maxGeneSetSize = 1e+10,
  print_genes = TRUE
)
}
\arguments{
\item{gene_list}{Named list object with gene IDs}

\item{gene_df}{Data frame including variable/module groups (column 1: group), and gene name (column2: gene). Can be used instead of gene_list}

\item{ID}{Character string for type of ID used in gene_list. One of SYMBOL, ENTREZ, ENSEMBL. Default is "SYMBOL"}

\item{species}{Character string denoting species of interest. "human" or "mouse" Default is "human"}

\item{collection}{Character string denoting Broad gene set database}

\item{subcollection}{Character string denoting Broad gene set sub-database}

\item{db}{Custom database}

\item{custom_bg}{Custom background. Formatted as a vector of gene IDs.}

\item{protein_coding}{TRUE or FALSE: do you want to limit the background to only protein-coding genes? Default is TRUE}

\item{minOverlap}{Minimum overlap between a gene set and your list of query genes for hypergeometric enrichment to be calculated. Default is 1.}

\item{minGeneSetSize}{Maximum overlap between a gene set and your list of query genes for hypergeometric enrichment to be calculated. Default is 10.}

\item{maxGeneSetSize}{Maximum size of a reference gene set for hypergeometric enrichment to be calculated. Default is 1e10}

\item{print_genes}{TRUE or FALSE. Do you want the results to include a list of genes that overlap between any given gene set and your query genes. Default is TRUE. Leaving this parameter as TRUE will make the function run slowly for very large datasets.}
}
\value{
Data frame enrichments including pathway, significance, and k/K ratios
}
\description{
Run hypergeometric enrichment on multiple lists using phyper. Permits gene symbols, Ensembl IDs, or Entrez IDs as gene identifiers and allows a background of protein-coding mouse genes or either whole-genome or protein-coding genes for human data. Also allows for custom backgrounds. Performs hypergeometric enrichment against a Broad Gene Set database or a custom database.
}
\examples{
gene_list <- list(HRV1 = names(example.gene.list[[1]]),
                  HRV2 = names(example.gene.list[[2]]))
flexEnrich(gene_list = gene_list, collection = "H", ID = "ENSEMBL")
}
\author{
Madison Cox
}
