% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_run.R
\name{GSEA_run}
\alias{GSEA_run}
\title{Run GSEA on multiple list using fgsea}
\usage{
GSEA_run(
  gene_list,
  gmt_ls = NULL,
  nperm = 1000,
  group = NULL,
  subgroup = NULL,
  S = "human"
)
}
\arguments{
\item{gene_list}{named list object with named numeric vectors of gene symbols and logFC}

\item{gmt_ls}{list object with gene ontology data}

\item{nperm}{number of permutations for P-value calculations. Default is 1000}

\item{group}{Which Geneset database you want to run GSEA on}

\item{subgroup}{Specify subgroup of Geneset DB you want to run GSEA on \cr
\tabular{rrrrr}{
\strong{Group} \tab    \strong{Sub-group}\cr
C1  \tab      \cr
C2  \tab      CGP\cr
C2  \tab      CP\cr
C2  \tab      CP:BIOCARTA\cr
C2  \tab      CP:KEGG\cr
C2  \tab      CP:PID\cr
C2  \tab      CP:REACTOME\cr
C2  \tab      CP:WIKIPATHWAYS\cr
C5  \tab      GO:BP\cr
C5  \tab      GO:CC\cr
C5  \tab      GO:MF\cr
C5  \tab      HPO\cr
C6  \tab     \cr
H   \tab      \cr
}}

\item{S}{Specify specie for Geneset database default is human}
}
\value{
Output of a list of list with GSEA results
}
\description{
Run GSEA on multiple list using fgsea
}
