% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterEnrich.R
\name{iterEnrich}
\alias{iterEnrich}
\title{Iterative Hypergeometric Enrichment}
\usage{
iterEnrich(
  anno_df = NULL,
  anno_featCol = NULL,
  anno_annotationCol = NULL,
  niter = 100,
  p_adjust = "fdr",
  ID = "SYMBOL",
  species = "human",
  collection = NULL,
  subcollection = NULL,
  db = NULL,
  custom_bg = NULL,
  protein_coding = TRUE,
  minOverlap = 1,
  minGeneSetSize = 10,
  maxGeneSetSize = 1e+10,
  print_genes = TRUE,
  ncores = 1
)
}
\arguments{
\item{anno_df}{Data frame containing features (such as methylation probes, SNP IDs, ATAC-seq peaks) and corresponding gene IDs for those features. Features with multiple annotations (ie rows) are iteratively sampled to 1 annotation per feature for enrichment}

\item{anno_featCol}{Character string. Name of the column containing the feature IDs (e.g. methylation probes, SNP IDs, ATAC-seq peaks). Default is the first column of anno_df}

\item{anno_annotationCol}{Character string. Name of the column containing the gene IDs (e.g. HGNC symbols, Entrez IDs, Ensembl IDs). Default is the second column of anno_df}

\item{niter}{Integer. Number of iterations. Default is 100.}

\item{p_adjust}{Character string. Method for p-value adjustment from p.adjust.methods(). Default is "fdr"}

\item{ID}{Character string. Type of gene annotation used in anno_df. One of SYMBOL, ENTREZ, ENSEMBL. Default is "SYMBOL"}

\item{species}{Character string denoting species of interest. "human" or "mouse." Default is "human"}

\item{collection}{Character string denoting Broad gene set database}

\item{subcollection}{Character string denoting Broad gene set sub-database
\tabular{rrrrr}{
\strong{collection} \tab    \strong{subcollection}\cr
C1  \tab      \cr
C2  \tab      CGP\cr
C2  \tab      CP\cr
C2  \tab      CP:BIOCARTA\cr
C2  \tab      CP:KEGG\cr
C2  \tab      CP:PID\cr
C2  \tab      CP:REACTOME\cr
C2  \tab      CP:WIKIPATHWAYS\cr
C3  \tab      \cr
C5  \tab      GO:BP\cr
C5  \tab      GO:CC\cr
C5  \tab      GO:MF\cr
C5  \tab      HPO\cr
C6  \tab      \cr
H   \tab      \cr
}}

\item{db}{If not using Broad databases, a data frame with gene ontology including gene set name (column 1: gs_name) and gene ID (column2: gene_symbol, entrez_gene, or ensembl_gene as matches your anno_df annotations)}

\item{custom_bg}{Custom background. Formatted as a vector of gene IDs.}

\item{protein_coding}{Logical. Do you want to limit the background to only protein-coding genes? Default is TRUE}

\item{minOverlap}{Minimum overlap between a gene set and your list of query genes for hypergeometric enrichment to be calculated. Default is 1. For the iterative function, the only valid value is 1 at the moment.}

\item{minGeneSetSize}{Maximum overlap between a gene set and your list of query genes for hypergeometric enrichment to be calculated. Default is 10.}

\item{maxGeneSetSize}{Maximum size of a reference gene set for hypergeometric enrichment to be calculated. Default is 1e10}

\item{print_genes}{Logical. Do you want the results to include a list of genes that overlap between any given gene set and your query genes. Default is FALSE. Setting this parameter as TRUE will make the function run slowly for very large datasets.}

\item{ncores}{Number of cores for parallel processing. Default is 1}
}
\value{
List of data frames includings
\itemize{
\item summary: median and min/max p-values, median k/K values and adjusted p-values. Note that medians treat iterations with no overlap in a gene set as p-value = 1 and k/K = 0. Gene lists reflect overlap between gene set and any annotation of a given feature in the input.
\item p_iterations: p-values for each iteration,
\item k/K_iterations:  k/K values for each iteration
}
}
\description{
Iteratively run flexEnrich on a list of features containing multiple annotations. Will select a random annotation for a given feature and run enrichment n times. Returns p-values and k/K ratios for each iteration, as well as a summative data frame with median p-values, median k/K ratios, and FDR-correction of the median p-values.
}
\examples{
# Each probe annoatated to multiple genes and each annotated to multiple probes
set.seed(42)
anno <- sample(names(example.gene.list[[1]])[1:20],size = 100, replace = TRUE)
df <- data.frame(feat = paste0("probe", rep(1:25, 4)), annotation = anno)

iterEnrich(anno_df = df,
           anno_featCol = "feat",
           anno_annotationCol = "annotation",
           niter = 3, ID = "ENSEMBL",
           collection = "H")
}
\author{
Madison Cox
}
