% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.human.db.pc.R
\docType{data}
\name{symbol.human.db.pc}
\alias{symbol.human.db.pc}
\title{Background genes for flexEnrich, HGNC symbols of protein-coding genes in the human genome}
\format{
A single-column matrix with gene symbols
\describe{
\item{x}{character. 19014 gene symbols.}
}
}
\source{
\url{http://ftp.ebi.ac.uk/pub/databases/genenames/hgnc/tsv/hgnc_complete_set.txt}
}
\usage{
symbol.human.db.pc
}
\description{
A single-column matrix of gene names
}
\keyword{datasets}
