% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.mouse.db.pc.R
\docType{data}
\name{symbol.mouse.db.pc}
\alias{symbol.mouse.db.pc}
\title{Background genes for flexEnrich, HGNC symbols of protein-coding genes in the mouse genome}
\format{
A single-column matrix with gene symbols
\describe{
\item{x}{character. 17682 gene symbols.}
}
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/mouse/collections.jsp}
}
\usage{
symbol.mouse.db.pc
}
\description{
A single-column matrix of gene names
}
\keyword{datasets}
