% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIGenrichr.R
\name{BIGenrichr}
\alias{BIGenrichr}
\title{Run EnrichR function on multiple Geneset databases on either gene list or modules of intertest}
\usage{
BIGenrichr(
  gene_mod_list,
  type = "gene",
  gene_id = "HGNC",
  dbs = c("MSigDB_Hallmark_2020"),
  mod_of_interest = "all"
)
}
\arguments{
\item{gene_mod_list}{Here input either Genelist in HGNC or Ensembl ID format or input module object output from module function}

\item{type}{Type can be either gene or module}

\item{gene_id}{gene id is either HGNC or Ensembl}

\item{dbs}{Enter the geneset database you want to run Enrichr on the default is MSigDB Hallmark 2020 to check available database run enrichR::listEnrichrDb()}

\item{mod_of_interest}{Input your modules of interest by default will run modules}
}
\value{
list with enrichr result for each geneset database
}
\description{
Run EnrichR function on multiple Geneset databases on either gene list or modules of intertest
}
