% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIGenrichr.R
\name{BIGenrichr}
\alias{BIGenrichr}
\title{Run EnrichR function on multiple Geneset databases on either gene list or modules of intertest}
\usage{
BIGenrichr(
  gene_list = NULL,
  gene_df = NULL,
  ID = "SYMBOL",
  dbs = c("MSigDB_Hallmark_2020")
)
}
\arguments{
\item{gene_list}{Here input either Genelist in SYMBOL or Ensembl ID format or input module object output from module function}

\item{gene_df}{Data frame including variable/module groups (column 1: group) and gene name (column2: gene). Can be used instead of gene_list}

\item{ID}{gene id is either SYMBOL, ENSEMBL, ENTREZ}

\item{dbs}{Enter the geneset database you want to run Enrichr on the default is MSigDB Hallmark 2020 to check available database run enrichR::listEnrichrDb()}
}
\value{
list with enrichr result for each geneset database
}
\description{
Run EnrichR function on multiple Geneset databases on either gene list or modules of intertest
}
\examples{
#Get gene names for enrichment
gene_list <- list(HRV1 = names(example.gene.list[[1]]),
                  HRV2 = names(example.gene.list[[2]]))
BIGenrichr(gene_list, ID = "ENSEMBL")

# Use gene_df
gene_df <- data.frame(gs_name = c(rep("HRV1", 100), rep("HRV2",100)),
                      gene = c(names(example.gene.list[[1]]),
                      names(example.gene.list[[2]])))
BIGenrichr(gene_df=gene_df, ID="ENSEMBL")
}
\author{
Basilin Benson
}
