% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIGsea.R
\name{BIGsea}
\alias{BIGsea}
\title{BIGsea}
\usage{
BIGsea(
  gene_list = NULL,
  gene_df = NULL,
  nperm = 1000,
  species = "human",
  ID = "SYMBOL",
  category = NULL,
  subcategory = NULL,
  db = NULL
)
}
\arguments{
\item{gene_list}{Named list object with named numeric vectors of gene symbols and logFC}

\item{gene_df}{Data frame including variable/module groups (column 1: group), gene name (column2: gene), and log fold change (column 3: logFC). Can be used instead of gene_list}

\item{nperm}{Numeric permutations for P-value calculations. Default is 1000}

\item{species}{Character string denoting species of interest. Default is "human"}

\item{ID}{Character string for type of ID used in gene_list. One of SYMBOL, ENTREZ, ENSEMBL. Default is "SYMBOL"}

\item{category}{Character string denoting Broad gene set database}

\item{subcategory}{Character string denoting Broad gene set sub-database. See https://www.gsea-msigdb.org/gsea/msigdb/}

\item{db}{If not using Broad databases, a data frame with gene ontology including gene set name (column 1: gs_name) and gene ID (column2: gene_symbol, entrez_gene, or ensembl_gene as matches your gene_list names)}
}
\value{
Data frame of enrichments including pathway, significance, and leading edge genes
}
\description{
Run gene set enrichment analysis (GSEA) on multiple list using fgsea. More information about GSEA and geneset databases check https://www.gsea-msigdb.org/gsea/msigdb
}
\examples{
BIGsea(example.gene.list, category="H", ID="ENSEMBL")
BIGsea(example.gene.list, category="C2", subcategory="CP", ID="ENSEMBL")

#Use gene_df
gene_df <- data.frame(gs_name = c("HRV1", rep("HRV2",100)),
                      gene = c("notReal",
                               names(example.gene.list[[2]])),
                     logFC = c(example.gene.list[[1]][1],
                               example.gene.list[[2]]))
BIGsea(gene_df=gene_df, category="H", ID="ENSEMBL")

#Use custom data base
db <- data.frame(module = c(rep("module1",10), rep("module2",10)),
                 symbol = sample(names(example.gene.list[[1]]), 20))
BIGsea(example.gene.list, ID="ENSEMBL", db=db)

}
