% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterateEnrich.R
\name{iterateEnrich}
\alias{iterateEnrich}
\title{Iterative Hypergeometric Enrichment}
\usage{
iterateEnrich(
  anno_df = NULL,
  anno_featCol = NULL,
  anno_annotationCol = NULL,
  niter = 100,
  p.adjust.method = "fdr",
  ID = "SYMBOL",
  species = "human",
  category = NULL,
  subcategory = NULL,
  db = NULL,
  custom_bg = NULL,
  protein_coding = TRUE,
  minOverlap = 1,
  minGeneSetSize = 10,
  maxGeneSetSize = 10000,
  print_genes = TRUE,
  ncores = 1
)
}
\arguments{
\item{anno_df}{A data frame containing features and corresponding gene IDs to subject to iterative hypergeometric enrichment}

\item{anno_featCol}{A string with the name of the column containing the feature IDs}

\item{anno_annotationCol}{A string with the name of the column containing the gene IDs}

\item{niter}{Number of iterations. Default is 100.}

\item{p.adjust.method}{Method for p-value adjustment from p.adjust.methods(). Default is "fdr"}

\item{ID}{Character string for type of ID used in gene_list. One of SYMBOL, ENTREZ, ENSEMBL. Default is "SYMBOL"}

\item{species}{Character string denoting species of interest. "human" or "mouse." Default is "human"}

\item{category}{Character string denoting Broad gene set database}

\item{subcategory}{Character string denoting Broad gene set sub-database}

\item{db}{Custom database}

\item{custom_bg}{Custom background. Formatted as a vector of gene IDs.}

\item{protein_coding}{TRUE or FALSE: do you want to limit the background to only protein-coding genes? Default is TRUE}

\item{minOverlap}{Minimum overlap between a gene set and your list of query genes for hypergeometric enrichment to be calculated. Default is 1. For the iterative function, the only valid value is 1 at the moment.}

\item{minGeneSetSize}{Maximum overlap between a gene set and your list of query genes for hypergeometric enrichment to be calculated. Default is 10.}

\item{maxGeneSetSize}{Maximum size of a reference gene set for hypergeometric enrichment to be calculated. Default is 1e10}

\item{print_genes}{TRUE or FALSE. Do you want the results to include a list of genes that overlap between any given gene set and your query genes. Default is TRUE. Leaving this parameter as TRUE will make the function run slowly for very large datasets.}

\item{ncores}{Number of cores for parallel processing. Default is 1}
}
\value{
List of data frames with length = number of gene sets: p-values for each iteration, k/K values for each iteration, summary with median p-values, k/K values, and adjusted p-values.
}
\description{
Iteratively run flexEnrich on a list of features containing multiple annotations. Will select a random annotation for a given feature and run enrichment n times. Returns p-values and k/K ratios for each iteration, as well as a summative data frame with median p-values, median k/K ratios, and FDR-correction of the median p-values.
}
\examples{
df <- data.frame("annotation" = names(example.gene.list[[1]]), "feat" = c(1:400))

iterateEnrich(anno_df = df, anno_featCol = "feat",
              anno_annotationCol = "annotation",
              niter = 5, ID = "ENSEMBL", category = "H")
}
\author{
Madison Cox
}
